Use double instead of float to store Float fields in MySQL
The precision of MySQL float is not enough to handle all Python floats.
From http://docs.python.org/tutorial/floatingpoint.html:
"...almost all platforms map Python floats to IEEE-754 “double precision”"

+++ trytond/backend/mysql/fields.py
@@ -55,7 +55,7 @@ class Float(fields.Float):
 
     @staticmethod
     def sql_type(field):
-        return ('float', 'float')
+        return ('double', 'double')
 
 
 class Numeric(fields.Numeric):
